requirements.txt
asgiref==3.7.2
Babel==2.12.1
Django==3.2.7
django-ckeditor==5.9.0
django-js-asset==2.1.0
django-modeltranslation==0.17.2
django-phonenumber-field==4.0.0
djangorestframework==3.11.1
phonenumbers==8.12.6
Pillow==10.0.0
pytz==2023.3
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.4
typing_extensions==4.7.1
